# apps/api/openapi/agent-tools.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pi Finance Agent HTTP contract",
    "version": "1.0.0"
  },
  "paths": {
...
    "/accounts": {
      "get": {
        "operationId": "list_accounts",
        "summary": "List accounts (active only by default; includeInactive also returns deactivated ones)",
        "x-pi-tool": {
          "label": "List Accounts",
          "result": {
            "kind": "items",
            "key": "accounts",
            "fields": {
              "id": "id",
              "name": "name",
              "kind": "kind",
              "balance_cents": "balanceCents",
              "active": {
                "path": "status",
                "equals": "active"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": true,
            "x-pi-context": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "context": false,
            "schema": {
              "type": "string",
              "enum": [
                "bank",
                "cash",
                "credit_card"
              ]
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "required": false,
            "x-pi-context": false,
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Opt-in: pass \"true\" to also return deactivated accounts (status inactive). Omitted or \"false\" keeps the active-only read. Workspace-scoped; credit cards are never included."
            }
          }
        ],
        "x-pi-tools": []
      },
      "post": {
        "operationId": "create_account",
        "summary": "HTTP contract for create_account",
        "x-pi-tool": {
          "name": "create_account",
          "label": "create_account",
          "description": "HTTP contract for create_account",
          "method": "POST",
          "path": "/accounts",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "bank",
                      "cash"
                    ]
                  },
                  "initialBalanceCents": {
                    "type": "integer",
                    "minimum": -1000000000000,
                    "maximum": 1000000000000
                  }
                },
                "required": [
                  "name",
                  "kind",
                  "initialBalanceCents"
                ]
              }
            }
          }
        }
      }
    },
...
    "/accounts/{accountId}/deactivate": {
      "post": {
        "operationId": "deactivate_account",
        "summary": "HTTP contract for deactivate_account",
        "x-pi-tool": {
          "name": "deactivate_account",
          "label": "deactivate_account",
          "description": "HTTP contract for deactivate_account",
          "method": "POST",
          "path": "/accounts/{accountId}/deactivate",
          "idempotency": true,
          "shadow": false,
          "result": null
        },
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
    "/accounts/{accountId}": {
      "get": {
        "summary": "Get one account",
        "operationId": "get_balance",
        "parameters": [
          {
            "name": "householdId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-pi-context": true
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-pi-tool": {
          "name": "get_balance",
          "label": "get_balance",
          "description": "HTTP contract for get_balance",
          "method": "GET",
          "path": "/accounts/{accountId}",
          "idempotency": false,
          "shadow": true,
          "result": null
        }
      },
      "patch": {
        "operationId": "update_account",
        "summary": "Update account",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "householdId",
            "in": "query",
            "required": false,
            "x-pi-context": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotencyKey",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  }
                },
                "required": []
              }
            }
          }
        },
        "x-pi-tool": {
          "name": "update_account",
          "idempotency": true
        }
      }
    },
...
  }
}